src/requirements.txt
Flask
flasgger
pyyaml
sklearn
gunicorn==20.1.0
nltk
pandas
prometheus-client